requirements.txt
amqp==5.2.0
appdirs==1.4.4
arrow==1.3.0
asgiref==3.7.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
beautifulsoup4==4.12.3
billiard==4.2.0
binaryornot==0.4.4
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
channels==4.0.0
channels-redis==4.2.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cookiecutter==2.5.0
cryptography==42.0.1
daphne==4.0.0
Django==5.0.1
django-browser-reload==1.12.1
django-tailwind==3.8.0
djangorestframework==3.14.0
frozendict==2.4.0
h11==0.14.0
html5lib==1.1
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
Jinja2==3.1.3
kombu==5.3.5
lxml==5.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.4
mdurl==0.1.2
msgpack==1.0.7
multitasking==0.0.11
numpy==1.26.2
outcome==1.3.0.post0
packaging==23.2
pandas==2.1.4
peewee==3.17.0
prompt-toolkit==3.0.43
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.17.2
pyOpenSSL==24.0.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-slugify==8.0.1
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
rich==13.7.0
selenium==4.16.0
service-identity==24.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
text-unidecode==1.3
trio==0.23.2
trio-websocket==0.11.1
Twisted==23.10.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
types-python-dateutil==2.8.19.20240106
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.13
webdriver-manager==4.0.1
webencodings==0.5.1
wsproto==1.2.0
yfinance==0.2.35
zope.interface==6.1
